% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actionButton_badge.R
\name{actionButton_badge}
\alias{actionButton_badge}
\title{Create a shiny actionButton with a badge}
\usage{
actionButton_badge(
  label,
  class = NULL,
  badge_color = "primary",
  badge_text = NULL,
  badge_class = NULL,
  badge_pill = FALSE,
  corner = FALSE,
  ...
)
}
\arguments{
\item{label}{label parameter of the actionButton}

\item{class}{class parameter of the actionButton}

\item{badge_color}{Color of the badge}

\item{badge_text}{Text of the badge}

\item{badge_class}{Additional HTML class for the button}

\item{badge_pill}{pill badge ? (TRUE or FALSE)}

\item{corner}{Badge in the corner ? (TRUE or FALSE)}

\item{...}{Others parameters of shiny actionButton}
}
\value{
an actionButton with a badge
}
\description{
Create a shiny actionButton with a badge
}
\examples{
actionButton_badge(label="Button",
inputId="btn_id",
badge_text = "5+",
badge_color = "success",
class="btn-primary",
corner = TRUE,
badge_pill = TRUE
)
}
