% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actionButton_circle.R
\name{actionButton_circle}
\alias{actionButton_circle}
\title{Create a shiny circular actionButton}
\usage{
actionButton_circle(style = NULL, icon = NULL, ...)
}
\arguments{
\item{style}{style parameter similar to actionButton}

\item{icon}{icon parameter similar to actionButton}

\item{...}{Same parameters as shiny::actionButton(...)}
}
\value{
A circular shiny actionButton
}
\description{
Create a shiny circular actionButton
}
\examples{
actionButton_circle(
inputId = "ok",
icon = shiny::icon("user"),
class="btn-success"
)
}
