% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actionButton_oc.R
\name{actionButton_oc}
\alias{actionButton_oc}
\title{Create a shiny actionButton that triggers an off canvas}
\usage{
actionButton_oc(id_oc, ...)
}
\arguments{
\item{id_oc}{id of the off canvas to trigger}

\item{...}{Same parameters as shiny::actionButton(...)}
}
\value{
A shiny actionButton that triggers a particular off canvas
}
\description{
Create a shiny actionButton that triggers an off canvas
}
\examples{
actionButton_oc(
inputId = "ok",
label = "Trigger", 
id_oc = "id_oc",
class="btn-danger"
)
}
