% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_badge.R
\name{add_badge}
\alias{add_badge}
\title{Create a BS5 badge}
\usage{
add_badge(color = "default", text = NULL, class = NULL, pill = FALSE)
}
\arguments{
\item{color}{Badge color}

\item{text}{Badge text}

\item{class}{Badge additional class}

\item{pill}{pill badge ? (TRUE or FALSE)}
}
\value{
a BS5 badge
}
\description{
Create a BS5 badge
}
\examples{
add_badge(
color ="transparent",
text = "25+",
class = "border border-danger text-danger",
pill = TRUE)
}
