% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_tooltip.R
\name{add_tooltip}
\alias{add_tooltip}
\title{Add a tooltip for a specific element}
\usage{
add_tooltip(
  tag = NULL,
  position = "top",
  trigger = "hover focus",
  text = NULL,
  color = "default"
)
}
\arguments{
\item{tag}{Element on which a tooltip will be added}

\item{position}{Tooltip position (top, bottom, left or right)}

\item{trigger}{Trigger of the tooltip (hover, focus or both)}

\item{text}{Tooltip message}

\item{color}{Tooltip color (black and white exclude)}
}
\value{
A tooltip for a specific element
}
\description{
Add a tooltip for a specific element
}
