% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/off_canvas.R
\name{off_canvas}
\alias{off_canvas}
\title{create a BS5 off canvas menu}
\usage{
off_canvas(id, title, html, position, close_btn = TRUE)
}
\arguments{
\item{id}{id of the off canvas menu}

\item{title}{HTML title of the off canvas menu}

\item{html}{HTML body of the off canvas menu}

\item{position}{position of the off-canvas (top, bottom, start or end)}

\item{close_btn}{close button (TRUE or FALSE)}
}
\value{
a BS5 off canvas menu
}
\description{
create a BS5 off canvas menu
}
\examples{
off_canvas("id","My off canvas menu","My content")
}
