% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_bs_color.R
\name{test_bs_color}
\alias{test_bs_color}
\title{Test the bootstrap validity of a color}
\usage{
test_bs_color(
  color,
  transparent = TRUE,
  default = FALSE,
  bw = TRUE,
  label = "Color"
)
}
\arguments{
\item{color}{color to test}

\item{transparent}{Is transparent color authorized ?}

\item{default}{Is default color authorized ?}

\item{bw}{Is black ans white authorized ?}

\item{label}{label for the error message}
}
\value{
A error message if the color is invalid
}
\description{
Test the bootstrap validity of a color
}
