% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_color.R
\name{test_hex_color}
\alias{test_hex_color}
\title{Test Hexcode colors}
\usage{
test_hex_color(x)
}
\arguments{
\item{x}{the colour value(s) to check.}
}
\value{
Logical vector indicating if strings(s) are valid hex representations
}
\description{
Check if string(s) are valid hex colour representation
}
