% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_addonbs5.R
\name{use_addonbs5}
\alias{use_addonbs5}
\title{Load tooltip and popover dependencies}
\usage{
use_addonbs5()
}
\value{
Include css dependencies of bootstrap 5 tooltip and popover
}
\description{
Load tooltip and popover dependencies
}
\examples{
use_addonbs5()
}
