% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_badge.R
\name{add_badge}
\alias{add_badge}
\title{Create a BS5 badge}
\usage{
add_badge(
  badge_color = "default",
  badge_text = NULL,
  badge_class = NULL,
  badge_pill = FALSE
)
}
\arguments{
\item{badge_color}{Badge color}

\item{badge_text}{Badge text}

\item{badge_class}{Badge additional class}

\item{badge_pill}{pill badge ? (TRUE or FALSE)}
}
\value{
a BS5 badge
}
\description{
Create a BS5 badge
}
\examples{
add_badge(
badge_color ="transparent",
badge_text = "25+",
badge_class = "border border-danger text-danger",
badge_pill = TRUE)
}
