% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_tooltip.R
\name{add_tooltip}
\alias{add_tooltip}
\title{Add a tooltip for a specific element}
\usage{
add_tooltip(
  tag = NULL,
  tooltip_position = "top",
  tooltip_trigger = "hover focus",
  tooltip_msg = NULL,
  tooltip_color = "default"
)
}
\arguments{
\item{tag}{Element on which a tooltip will be added}

\item{tooltip_position}{Tooltip position (top, bottom, left or right)}

\item{tooltip_trigger}{Trigger of the tooltip (hover, focus or both)}

\item{tooltip_msg}{Tooltip message}

\item{tooltip_color}{Tooltip color (black and white exclude)}
}
\value{
A tooltip for a specific element
}
\description{
Add a tooltip for a specific element
}
